requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
pydantic==2.6.4
httpx==0.28.1
requests==2.32.4

streamlit==1.46.1

python-dotenv==1.1.1

openai==1.93.1

loguru==0.7.2
typing-extensions==4.14.1
anyio==4.9.0
pytest==8.2.1
pytest-asyncio==0.23.6

